OK, this actually *is* working...
At least partly. Faults are found in all of the documents in the activitystreams-test-documents suite, but this is because

1. five have no `@context` value;
2. the remaining two hundred all consistently use 'http' as the protocol part of the `@context` URI, but the specification requires `https`...

I am picking up other faults and some (perhaps all) of these may be errors in my code rather than in the test documents, but I am heading in the right direction.